+++ kinto/core/authorization.py
@@ -221,6 +221,9 @@ class RouteFactory(object):
         method = request.method.lower()
         required_permission = self.method_permissions.get(method)
 
+        # For create permission, the object id is the plural endpoint.
+        collection_path = service.collection_path.format(**request.matchdict)
+
         # In the case of a "PUT", check if the targetted record already
         # exists, return "write" if it does, "create" otherwise.
         if request.method.lower() == "put":
@@ -233,10 +236,15 @@ class RouteFactory(object):
             except storage_exceptions.RecordNotFoundError:
                 # The record does not exist, the permission to create on
                 # the related collection is required.
-                permission_object_id = service.collection_path.format(
-                    **request.matchdict)
+                permission_object_id = collection_path
                 required_permission = "create"
             else:
-                required_permission = "write"
+                # For safe creations, the user needs a create permission.
+                # See Kinto/kinto#792
+                if request.headers.get('If-None-Match') == '*':
+                    permission_object_id = collection_path
+                    required_permission = "create"
+                else:
+                    required_permission = "write"
 
         return (permission_object_id, required_permission)